Handle being called between curly braces in completeMapping
Previously if miniSnip#completeMapping() was called with a curly brace on
the right the following error would occur:
```
Error detected while processing function miniSnip#completeMapping[12]..miniSnip#completeFunc:
line   14:
E219: Missing {.
```

The issue was the miniSnip#completeMapping() did not remove the curly
brace when calling for completion. This commit solves this by using
regex to get only the text from the cursor to the first alphabetic
character aka `\a`.

I chose `\a` since that was what ws previously used
to determine the start of the word.